The steps should be whole seconds

+++ video.php
@@ -83,8 +83,7 @@ class video
 		    throw new InvalidArgumentException('Steps is not numeric');
 		if($duration<$steps)
 			throw new InvalidArgumentException(sprintf('File duration is %d, not able to make %d snapshots', $duration, $steps));
-		$step=floor($duration/($steps+1)); //Get the step size
-		
+		$step=(int)floor($duration/($steps+1)); //Get the step size
 		$step_list=range($step,$duration,$step); //Make an array with the positions
 		if($first!==false)
 			array_unshift($step_list,1);
